Extract `timed` function utility
